modified:   contracts/exchange/ExchangeDomain.sol 	modified:   contracts/exchange/ExchangeOrdersHolder.sol 	modified:   contracts/exchange/ExchangeState.sol 	modified:   contracts/exchange/NftExchange.sol

+++ contracts/exchange/ExchangeDomain.sol
@@ -2,14 +2,18 @@
 pragma solidity >=0.4.22 <0.9.0;
 pragma experimental ABIEncoderV2;
 
-
 contract ExchangeDomain {
-
-    enum AssetType {ETH, ERC20, ERC1155, ERC721, ERC721Deprecated}
+    enum AssetType {
+        ETH,
+        ERC20,
+        ERC1155,
+        ERC721,
+        ERC721Deprecated
+    }
 
     struct Asset {
         address token;
-        uint tokenId;
+        uint256 tokenId;
         AssetType assetType;
     }
 
@@ -17,25 +21,21 @@ contract ExchangeDomain {
         /* who signed the order */
         address owner;
         /* random number */
-        uint salt;
-
+        uint256 salt;
         /* what has owner */
         Asset sellAsset;
-
         /* what wants owner */
         Asset buyAsset;
     }
 
     struct Order {
         OrderKey key;
-
         /* how much has owner (in wei, or UINT256_MAX if ERC-721) */
-        uint selling;
+        uint256 selling;
         /* how much wants owner (in wei, or UINT256_MAX if ERC-721) */
-        uint buying;
-
+        uint256 buying;
         /* fee for selling */
-        uint sellerFee;
+        uint256 sellerFee;
     }
 
     /* An ECDSA signature. */
